requirements.txt
# Autonomous Market Intelligence Agent — dependencies
# Python 3.10+ recommended

# HTTP & feeds
requests>=2.28.0
feedparser>=6.0.0

# Config & env
pyyaml>=6.0
python-dotenv>=1.0.0

# LLM (pick one or use litellm for multi-provider)
openai>=1.0.0
# anthropic>=0.18.0

# Optional: orchestration
# langgraph>=0.2.0
# langchain-core>=0.2.0

# Optional: DB
# (stdlib sqlite3 is enough; for async or PostgreSQL add relevant driver)
